feat: create sale improvements
