+++ package.json
@@ -34,6 +34,7 @@
     "keccak256": "^1.0.6",
     "merkletreejs": "^0.4.0",
     "mixpanel-browser": "^2.47.0",
+    "moment": "^2.30.1",
     "next": "^14.1.3",
     "react": "^18.2.0",
     "react-cookie": "^7.1.0",
